+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+-   Add typings.
+-   **Breaking:** The list of props is now passed to `makeWebComponents` as the second argument instead of being a property on the function.
+
 ## [0.0.4]
 
 ### Fixed